:sparkles: feat: Amélioration de l'API des épisodes
Ajoutez une nouvelle fonctionnalité pour sauvegarder un épisode dans la base de données avant de l'ajouter à la liste de lecture. Modifiez également le type des ID TMDB dans SeasonEntity de int à Long et le numéro de la saison de int à Integer. Ajoutez une nouvelle exception EpisodeException.

+++ domain/src/main/java/org/lafabrique_epita/domain/entities/SeasonEntity.java
@@ -25,11 +25,11 @@ public class SeasonEntity extends MasterClass {
     private String overview;
 
     @Column(nullable = false, unique = true)
-    private int idTmdb;
+    private Long idTmdb;
 
     private String posterPath;
 
-    private int seasonNumber;
+    private Integer seasonNumber;
 
 
     @ManyToOne(optional = false)
